reorganize existing and new sqlite tests; consolidate sqlite fixtures
