+++ pycytominer/tests/test_cyto_utils/sqlite/test_sqlite_clean.py
@@ -1,10 +1,6 @@
 """ 
-Tests for:
-pycytominer.cyto_utils.sqlite.clean
-pycytominer.cyto_utils.sqlite.meta
+Tests for: pycytominer.cyto_utils.sqlite.clean
 """
-
-import tempfile
 
 import pytest
 from pycytominer.cyto_utils.sqlite.clean import (
@@ -14,116 +10,9 @@ from pycytominer.cyto_utils.sqlite.clean import (
     update_columns_to_nullable,
     update_values_like_null_to_null,
 )
-from pycytominer.cyto_utils.sqlite.meta import (
-    LIKE_NULLS,
-    collect_columns,
-    engine_from_str,
-)
-from sqlalchemy import create_engine
+from pycytominer.cyto_utils.sqlite.meta import LIKE_NULLS
 from sqlalchemy.engine.base import Engine
 from sqlalchemy.exc import IntegrityError
-
-
-@pytest.fixture
-def database_engine_for_testing() -> Engine:
-    """
-    A database engine for testing as a fixture to be passed
-    to other tests within this file.
-    """
-
-    # get temporary directory
-    tmpdir = tempfile.gettempdir()
-
-    # create a temporary sqlite connection
-    sql_path = f"sqlite:///{tmpdir}/test_sqlite.sqlite"
-
-    engine = create_engine(sql_path)
-
-    # statements for creating database with simple structure
-    create_stmts = [
-        "drop table if exists tbl_a;",
-        """
-        create table tbl_a (
-        col_integer INTEGER NOT NULL
-        ,col_text TEXT
-        ,col_blob BLOB
-        ,col_real REAL
-        );
-        """,
-        "drop table if exists tbl_b;",
-        """
-        create table tbl_b (
-        col_integer INTEGER
-        ,col_text TEXT
-        ,col_blob BLOB
-        ,col_real REAL
-        );
-        """,
-    ]
-
-    insert_vals = [1, "sample", b"sample_blob", 0.5]
-
-    with engine.begin() as connection:
-        for stmt in create_stmts:
-            connection.execute(stmt)
-
-        # insert statement with some simple values
-        # note: we use SQLAlchemy's parameters to insert data properly, esp. BLOB
-        connection.execute(
-            (
-                "INSERT INTO tbl_a (col_integer, col_text, col_blob, col_real)"
-                "VALUES (?, ?, ?, ?);"
-            ),
-            insert_vals,
-        )
-        connection.execute(
-            (
-                "INSERT INTO tbl_b (col_integer, col_text, col_blob, col_real)"
-                "VALUES (?, ?, ?, ?);"
-            ),
-            insert_vals,
-        )
-
-    return engine
-
-
-def test_engine_from_str():
-    """
-    Testing engine_from_str
-    """
-    # test str functionality
-    engine = engine_from_str(":memory:")
-    assert isinstance(engine, Engine)
-    assert engine.execute("PRAGMA integrity_check").fetchone()[0] == "ok"
-
-    # test sqlalchemy engine
-    engine = engine_from_str(create_engine("sqlite:///:memory:"))
-    assert isinstance(engine, Engine)
-    assert engine.execute("PRAGMA integrity_check").fetchone()[0] == "ok"
-
-
-def test_collect_columns(database_engine_for_testing):
-    """
-    Testing collect_columns
-    """
-
-    # test full database columns collected
-    assert len(collect_columns(database_engine_for_testing)) == 8
-
-    # test single database table collected
-    assert collect_columns(database_engine_for_testing, table_name="tbl_a") == [
-        ("tbl_a", "col_integer", "INTEGER", 1),
-        ("tbl_a", "col_text", "TEXT", 0),
-        ("tbl_a", "col_blob", "BLOB", 0),
-        ("tbl_a", "col_real", "REAL", 0),
-    ]
-
-    # test single column from single table collected
-    assert collect_columns(
-        database_engine_for_testing,
-        table_name="tbl_b",
-        column_name="col_integer",
-    ) == [("tbl_b", "col_integer", "INTEGER", 0)]
 
 
 def test_contains_conflicting_aff_storage_class(database_engine_for_testing):
